Fix terminfo
Turns out by default tmux-256color is not available, so I added a script to generate it in ~/.terminfo. Then we configure alacritty to use xterm-256, because otherwise it reports Alacritty, and then I have to use that in the tmux conf. That in turn makes tmux.conf not compatible with iTerm and Terminal.app.

So I settled on Alacritty reporting xterm-256color, then tmux uses tmux-256color inside and forwards to xterm-256color:RGB outside. This way true color works properly in and out.